requirements.txt
# Core ML Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6

# MLflow for Experiment Tracking
mlflow==2.8.1

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Kafka Client
kafka-python==2.0.2

# Redis for Feature Caching
redis==5.0.1

# Data Validation and Drift Detection
scipy==1.11.3

# Logging and Monitoring
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
